An experiment to make a nice index page for cargo central.
The index is a "gh_pages" github page. Usually you would use an orphaned branch
for the page, but since the json used to build the index is in master, gh_pages is based on master,
and should be kept up to date by merging master in when there are changes.